fix(miopen/tensor): widen offset param of SetTensor/ScaleTensor to size_t
The offset parameter of SetTensor and ScaleTensor was declared as int
in both the host function signature (tensor_ops.hpp) and the GPU kernel
(MIOpenSubTensorOpWithScalarKernel.cpp). Callers in rnnocl.cpp pass
size_t values for this offset — e.g. hid_shift = li * batch_n *
hy_stride which reaches 2,764,800,000 for layer 3 of the LargeWorkspace
LSTM config (> INT_MAX = 2,147,483,647). The implicit size_t -> int
narrowing at the call site truncated the offset to a negative value.
In the kernel, dst[i + offset] then computed a wrapped unsigned address
pointing outside the allocated buffer, causing the GPU page fault.

Fix: widen offset to size_t in the host signatures and unsigned int in
the GPU kernel parameters. The existing stride/len params were widened
to unsigned int in the previous commit; this commit completes the fix
by covering the offset path that is the actual crash trigger.

Fixes Standard/GPU_LSTM_LargeWorkspace_FP16.ForwardInferenceSucceedsWhenWorkspaceExceedsInt/0

+++ projects/miopen/src/include/miopen/tensor_ops.hpp
@@ -145,13 +145,13 @@ MIOPEN_INTERNALS_EXPORT void ScaleTensor(const Handle& handle,
                                          const TensorDescriptor& yDesc,
                                          Data_t y,
                                          const void* alpha,
-                                         int offset = 0);
+                                         size_t offset = 0);
 
 MIOPEN_INTERNALS_EXPORT void SetTensor(const Handle& handle,
                                        const TensorDescriptor& yDesc,
                                        Data_t y,
                                        const void* alpha,
-                                       int offset = 0);
+                                       size_t offset = 0);
 
 MIOPEN_INTERNALS_EXPORT void OpTensor(const Handle& handle,
                                       miopenTensorOp_t tensorOp,
